Generalizing value of STAT arguments
Generalizing value of STAT arguments should be 5 after sorting and taking out uniq values instead of 6 somewhere and 7 somewhere else. In Amazon AWS AMI the value was 7 because of: 
[root@ip-172-31-16-92 ~]# df -aP |grep \/opt\/gluu-server-3.0.2\/ | awk '{ print $6 }' | grep -Eohw 'proc|lo|pts|modules|dev'
dev
proc
lo
modules
dev
pts
proc
So total above is 7 and hence our startup was a problem.

Now, we're using: 
[root@ip-172-31-16-92 ~]# df -aP |grep \/opt\/gluu-server-3.0.2\/ | awk '{ print $6 }' | grep -Eohw 'proc|lo|pts|modules|dev'|sort|uniq
dev
lo
modules
proc
pts
Now the value is universal of all systems.